requirements.txt
distro==1.9.0
JPype1==1.5.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tabula-py==2.10.0
tzdata==2024.2
